+++ restapi/server/operations/medco_node/get_cohorts.go
@@ -112,7 +112,7 @@ func (o *GetCohortsDefaultBody) UnmarshalBinary(b []byte) error {
 type GetCohortsOKBodyItems0 struct {
 
 	// cohort Id
-	CohortID float64 `json:"cohortId,omitempty"`
+	CohortID int64 `json:"cohortId,omitempty"`
 
 	// cohort name
 	CohortName string `json:"cohortName,omitempty"`
@@ -121,7 +121,7 @@ type GetCohortsOKBodyItems0 struct {
 	CreationDate string `json:"creationDate,omitempty"`
 
 	// query Id
-	QueryID float64 `json:"queryId,omitempty"`
+	QueryID int64 `json:"queryId,omitempty"`
 
 	// update date
 	UpdateDate string `json:"updateDate,omitempty"`
